Improve search-word plot for non-english languages.
Every search-query is prepended with 'Searched for XY', (I really don't
know why YouTube is doing this but whatever). However this prefix is
depending on the language your google account is set to. In german for
example it is 'Nach XY gesucht'. This bug got fixed in this commit and
should now work with data in any language.