Start converting the visualization into a Telegrambot.
At this point there is a database with users, the users can set their
timezone and the bot can visualize the the search data.

However, the bot cannot yet visualize the watch data (but shouldn't be
too hard since most of the code is already written just not in the bot)

Whatever, I am going to sleep now ... (Really should have done this
hours ago).